Installer for any Software
This includes an exported exe file as well as the logo for the app (not added in the .exe but you can use it in a shortcut to make it look nice).

The Ver 1.1.exe is an installer that will install the programm with all the dependencies.

Next updates will include the option to have direct access to the Data Base for the people who know how to use a DB and would like to tinker with their own files.